# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transparência API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/taxas": {
            "get": {
                "tags": [
                    "/taxas"
                ],
                "summary": "listar taxas",
                "operationId": "c545976d151cf628a631bf21bcd0a630",
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/taxas"
                ],
                "summary": "registrar taxa",
                "operationId": "2b99392547b0497501c2dbcb55bd4101",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "c_desctaxa",
                                    "n_valotaxa",
                                    "n_permtaxa",
                                    "d_denvtaxa",
                                    "c_freqtaxa"
                                ],
                                "properties": {
                                    "c_desctaxa": {
                                        "description": "descricão da taxa",
                                        "type": "string"
                                    },
                                    "n_valotaxa": {
                                        "description": "valor da taxa",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "n_vmultaxa": {
                                        "description": "valor da multa",
                                        "type": "integer"
                                    },
                                    "n_permtaxa": {
                                        "description": "valor da multa, percentagem",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "d_denvtaxa": {
                                        "description": "data de início de envio da taxa",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "c_freqtaxa": {
                                        "description": "frequência de envio da taxa",
                                        "type": "string"
                                    },
                                    "n_praztaxa": {
                                        "description": "prazo em dias, para o pagamento da taxa",
                                        "type": "integer"
                                    },
                                    "n_codicoord": {
                                        "description": "id do coordenador do que criou a taxa",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "taxa registrada com sucesso"
                    },
                    "412": {
                        "description": "Erro ao validar os dados"
                    },
                    "404": {
                        "description": "coordenador não encontrado"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/taxas/{taxa}": {
            "get": {
                "tags": [
                    "/taxas"
                ],
                "summary": "mostrar um Taxa",
                "operationId": "37f65b249240225ccad63ca24d6bc9eb",
                "parameters": [
                    {
                        "name": "taxa",
                        "in": "path",
                        "description": "id do taxa",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "404": {
                        "description": "Taxa não encontrado"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/taxas"
                ],
                "summary": "atualizar taxa",
                "operationId": "00779debabe2819bc419495c2db68a88",
                "parameters": [
                    {
                        "name": "taxa",
                        "in": "path",
                        "description": "id do taxa",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "c_desctaxa": {
                                        "description": "descricão da taxa",
                                        "type": "string"
                                    },
                                    "n_valotaxa": {
                                        "description": "valor da taxa",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "n_vmultaxa": {
                                        "description": "valor da multa",
                                        "type": "integer"
                                    },
                                    "n_permtaxa": {
                                        "description": "valor da multa, percentagem",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "d_denvtaxa": {
                                        "description": "data de início de envio da taxa",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "c_freqtaxa": {
                                        "description": "frequência de envio da taxa",
                                        "type": "string"
                                    },
                                    "n_praztaxa": {
                                        "description": "prazo em dias, para o pagamento da taxa",
                                        "type": "integer"
                                    },
                                    "n_codicoord": {
                                        "description": "id do coordenador do que criou a taxa",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "taxa registrada com sucesso"
                    },
                    "412": {
                        "description": "Erro ao validar os dados"
                    },
                    "404": {
                        "description": "coordenador não encontrado"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/taxas"
                ],
                "summary": "deletar taxa",
                "operationId": "2ff5ccb36bc9e9d7876dff09e4c2f694",
                "parameters": [
                    {
                        "name": "taxa",
                        "in": "path",
                        "description": "id da taxa",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "taxa deletada com sucesso!"
                    },
                    "404": {
                        "description": "taxa não encontrada"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}